Voice Search Icon Reorganization (#1476)
I reorganized the voice search icon logic to clean things up

I also noticed a bug when both the mic and the listening icon are custom. The styling changes fix that bug

J=none
TEST=manual

Test that the voice icon works and that all combinations of custom icons works as well
